modified sample according to new changes
