--- testdata/memcached-quarkus-operator/k8s/deployment.yaml
@@ -1,58 +1,0 @@
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: memcached-operator
-
----
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: memcached-operator
-  namespace: memcached-operator
-
----
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: memcached-operator
-  namespace: memcached-operator
-spec:
-  selector:
-    matchLabels:
-      app: memcached-operator
-  template:
-    metadata:
-      labels:
-        app: memcached-operator
-    spec:
-      serviceAccountName: memcached-operator
-      containers:
-      - name: operator
-        image: eu.gcr.io/adamsandor-test/memcached-operator
-        imagePullPolicy: Always
-        ports:
-        - containerPort: 80
-        readinessProbe:
-          httpGet:
-            path: /health
-            port: 8080
-          initialDelaySeconds: 1
-        livenessProbe:
-          httpGet:
-            path: /health
-            port: 8080
-          initialDelaySeconds: 30
-
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  name: memcached-operator-admin
-subjects:
-- kind: ServiceAccount
-  name: memcached-operator
-  namespace: memcached-operator
-roleRef:
-  kind: ClusterRole
-  name: cluster-admin
-  apiGroup: ""
