+++ testdata/memcached-quarkus-operator/src/main/java/com/example/Memcached.java
@@ -1,12 +1,11 @@
-package com.lucky;
+package com.example;
 
 import io.fabric8.kubernetes.api.model.Namespaced;
 import io.fabric8.kubernetes.client.CustomResource;
 import io.fabric8.kubernetes.model.annotation.Group;
 import io.fabric8.kubernetes.model.annotation.Version;
 
-@Version("v1alpha1")
+@Version("v1")
 @Group("cache.example.com")
-public class Memcached extends CustomResource<MemcachedSpec, MemcachedStatus>
-    implements Namespaced {}
+public class Memcached extends CustomResource<MemcachedSpec, MemcachedStatus> implements Namespaced {}
