+++ testdata/memcached-quarkus-operator/src/main/java/com/example/MemcachedSpec.java
@@ -1,9 +1,9 @@
-package com.lucky;
+package com.example;
 
 public class MemcachedSpec {
 
     // Add Spec information here
-
+    // Size is the size of the memcached deployment
     private Integer size;
 
     public Integer getSize() {
