--- testdata/memcached-quarkus-operator/src/main/java/com/lucky/JavaOpOperator.java
@@ -1,26 +1,0 @@
-
-package com.lucky;
-
-import io.javaoperatorsdk.operator.Operator;
-import io.quarkus.runtime.Quarkus;
-import io.quarkus.runtime.QuarkusApplication;
-import io.quarkus.runtime.annotations.QuarkusMain;
-import javax.inject.Inject;
-
-@QuarkusMain
-public class JavaOpOperator implements QuarkusApplication {
-
-  @Inject Operator operator;
-
-  public static void main(String... args) {
-    Quarkus.run(JavaOpOperator.class, args);
-  }
-
-  @Override
-  public int run(String... args) throws Exception {
-    operator.start();
-
-    Quarkus.waitForExit();
-    return 0;
-  }
-}
