+++ testdata/memcached-quarkus-operator/src/main/resources/application.properties
@@ -1,3 +1,7 @@
 quarkus.container-image.build=true
 #quarkus.container-image.group=
-quarkus.container-image.name=java-op-service
+quarkus.container-image.name=memcached-quarkus-operator-operator
+# set to true to automatically apply CRDs to the cluster when they get regenerated
+quarkus.operator-sdk.crd.apply=false
+# set to true to automatically generate CSV from your code
+quarkus.operator-sdk.generate-csv=false
