+++ testdata/memcached-quarkus-operator/src/main/resources/memcached-sample.yaml
@@ -1,7 +1,7 @@
-apiVersion: cache.example.com/v1alpha1
+apiVersion: cache.example.com/v1
 kind: Memcached
 metadata:
   name: memcached-sample
 spec:
   # Add fields here
-  size: 5
+  size: 1
